scripts/requirements.txt
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
requests>=2.28.0
